Lower Requirements Improve Performance
Lowered the values of some items to make the world smaller to improve
efficiency. World size was 1000000 which was unnecessarily large
(probably 90% was just trees in darkness...). Also reduced the skybox
image from something like 50000x2000 to 1000x600 as it works just as
well at cuts the file size from 11mb to 120kb. Reducing the world size
also means reducing the number of trees so no longer processing all of
them.