Modified Comment
Updated comment on mountain to reflect the change in code. In the normal
Diamond-Square algorithm the four corners are averaged and then an error
added to calculate the offset. I've modified to include a coord from the
centre as well, this makes it look more like a mountain rather than just
a mountainy terrain.